edit create employee by adding cretion of pin automaticly, replace the signature pad feature with the find employee by the pin for the tool movements

+++ src/app/api/tools/[id]/route.ts
@@ -19,16 +19,17 @@ export async function GET(
     }
 
     const { ROLE, POSITION } = tokenData;
+    const isSuperAdmin = ROLE === Role.SUPER_ADMIN;
+    const isAdminStoreKeeper =
+      ROLE === Role.ADMIN && POSITION === Position.STORE_KEEPER;
 
-    if (
-      (ROLE !== Role.SUPER_ADMIN && POSITION !== Position.MANAGER) ||
-      (ROLE !== Role.ADMIN && POSITION !== Position.STORE_KEEPER)
-    ) {
+    if (!isSuperAdmin && !isAdminStoreKeeper) {
       return NextResponse.json(
-        { message: "You are Not the Super Admin or the Admin🤨" },
+        { message: "You are Not the Super Admin or the Storekeeper Admin🤨" },
         { status: 403 }
       );
     }
+
     const { id } = await params;
     await connectToDB();
 
@@ -53,15 +54,14 @@ export async function PATCH(
         { status: 401 }
       );
     }
-
     const { ROLE, POSITION } = tokenData;
+    const isSuperAdmin = ROLE === Role.SUPER_ADMIN;
+    const isAdminStoreKeeper =
+      ROLE === Role.ADMIN && POSITION === Position.STORE_KEEPER;
 
-    if (
-      (ROLE !== Role.SUPER_ADMIN && POSITION !== Position.MANAGER) ||
-      (ROLE !== Role.ADMIN && POSITION !== Position.STORE_KEEPER)
-    ) {
+    if (!isSuperAdmin && !isAdminStoreKeeper) {
       return NextResponse.json(
-        { message: "You are Not the Super Admin or the Admin🤨" },
+        { message: "You are Not the Super Admin or the Storekeeper Admin🤨" },
         { status: 403 }
       );
     }
@@ -86,6 +86,26 @@ export async function DELETE(
   { params }: { params: Promise<{ id: string }> }
 ) {
   try {
+    const tokenData = await checkToken();
+
+    if (!tokenData) {
+      return NextResponse.json(
+        { message: "Invalid token or not authenticated" },
+        { status: 401 }
+      );
+    }
+
+    const { ROLE, POSITION } = tokenData;
+    const isSuperAdmin = ROLE === Role.SUPER_ADMIN;
+    const isAdminStoreKeeper =
+      ROLE === Role.ADMIN && POSITION === Position.STORE_KEEPER;
+
+    if (!isSuperAdmin && !isAdminStoreKeeper) {
+      return NextResponse.json(
+        { message: "You are Not the Super Admin or the Storekeeper Admin🤨" },
+        { status: 403 }
+      );
+    }
     const { id } = await params;
     await connectToDB();
     const tool = await Tool.findByIdAndDelete(id);
